% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPlay.R
\name{autoPlay}
\alias{autoPlay}
\title{Play Audio without an Explicit Play Button}
\usage{
autoPlay(audioId, src)
}
\arguments{
\item{audioId}{character. The id of the audio element.}

\item{src}{character. The name of a WAV file in the www folder.}
}
\value{
When included in a reactive observer (\code{observe()} or \code{observeEvent()}), plays an audio file without a click from a participant.
}
\description{
Play Audio without an Explicit Play Button
}
\examples{
# NOTE: This example will only play audio from a saved app script in the same directory as a 'www' subdirectory that contains the audio file.
data("samp_wav")
tuneR::writeWave("samp_wav", "www/sample.wav")

library(shiny)

ui <- tabsetPanel(id = "tabs",
                  tabPanel(
                    "tab1",
                    "The sound is not playing yet. Click on tab2 to play the sound."
                  ),
                  tabPanel(
                    "tab2",
                    "The sound should be playing now."
                  )
)

server <- function(input, output, session) {
  observeEvent(input$tabs, {
    if (input$tabs == "tab2") {
      autoPlay(audioId = "sample", src = "sample.wav")
    }
  })
}

}
