% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cal_noise}
\alias{cal_noise}
\title{A sample of white noise}
\format{
A Wave object (for more information, see \code{tuneR::readWave()} documentation) containing a 10 second sample of white noise sampled at 44.1 kHz.
}
\source{
\url{https://github.com/ChaitLabUCL/HeadphoneCheck_Test}
}
\usage{
cal_noise
}
\description{
A 10 second sample of white noise useful for presenting to the user for headphone volume calibration in a shiny app.
}
\keyword{datasets}
