% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2hex.R
\name{col2hex}
\alias{col2hex}
\title{Convert color name to hexidecimal code}
\usage{
col2hex(cname)
}
\arguments{
\item{cname}{Character. Name of an HTML color.}
}
\value{
A hexidecimal color code for the provided color name.
}
\description{
Convert color name to hexidecimal code
}
\details{
Pulled from the \code{gplots} package. This version allows for compatibility with older versions of R.
}
\examples{
col2hex("white")

}
