% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consentServer.R
\name{consentServer}
\alias{consentServer}
\title{Shiny Server function for documenting participant consent}
\usage{
consentServer(
  id = "consent",
  trigger,
  result = c("disable", "hide"),
  delayResponse = 1,
  agreeLab = "Agree",
  disagreeLab = NULL,
  cons2rec = FALSE,
  cons2recRequire = TRUE,
 
    cons2recLab = "This experiment requires us to make audio recordings of your voice. Do you consent to having your voice recorded?",
  cons2recYes = "I consent to my voice being recorded for this research.",
  cons2recNo = "I do NOT grant my consent for my voice to be recorded."
)
}
\arguments{
\item{id}{The input ID associated with the consent module. Must match the ID of \code{consentUI()}.}

\item{trigger}{A reactive value indicating the event that should trigger the appearance of the consent form. May be an \code{input$...} value from outside the module wrapped in \code{reactive()}.}

\item{result}{A character value describing what should happen when the participant gives consent. Must be either "disable" or "hide".}

\item{delayResponse}{Integer. How long (in milliseconds) should the participant be forced to wait (i.e., to read the form text) before clicking "Agree" or "Disagree"? Defaults to no delay.}

\item{agreeLab}{Character. The text to be displayed on the button that the participant will click to give their consent. Defaults to "I agree".}

\item{disagreeLab}{If not \code{NULL} (the default), a button will be shown alongside the "agree" button, that the participant may click to decline their consent. The character string provided here will serve as the label for that button.}

\item{cons2rec}{Boolean. Should the participant be required to explicitly give their consent to be recorded? Default is \code{FALSE}.}

\item{cons2recRequire}{Boolean. If a separate consent to record is shown, should the participant be required to consent to voice recording before the can agree to the rest of the consent form and continue with the experiment? Defaults to \code{TRUE}.}

\item{cons2recLab}{Character. Label for the "Consent to Record" radio button pane. Defaults to a generic question about permission to record audio.}

\item{cons2recYes}{Character. Label for the radio button the participant will click to agree to audio recording. Defaults to a generic statement.}

\item{cons2recNo}{Character. Label for the radio button the participant will click to decline consent to audio recording. Defaults to a generic statement.}
}
\value{
A reactive value of type \code{integer} indicating whether the participant has given consent to the experiment and to record (=2), consent to everything not including recording (=1), or not consented to any portion of the experiment (=0).
}
\description{
Shiny Server function for documenting participant consent
}
\examples{
if (interactive()) {
ui <- shiny::fluidPage(
  consentUI(id = "cons")
)

server <- function(input, output, session) {
 consent <- consentServer(id = "cons")
}
shiny::shinyApp(ui = ui, server = server)
}


}
\seealso{
Must be used with \code{\link{consentUI}}. If \code{cons2rec = TRUE}, consider using \code{\link{recorderUI}} and \code{\link{recorderServer}}.

Other Consent form module: 
\code{\link{consentUI}()}
}
\concept{Consent form module}
