% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consentUI.R
\name{consentUI}
\alias{consentUI}
\title{Shiny UI function for displaying a participant consent form}
\usage{
consentUI(id = "consent", title = "Consent Form", filename = NULL)
}
\arguments{
\item{id}{The input ID associated with the consent module. Must match the ID of \code{consentServer()}.}

\item{title}{The title of the consent form. Defaults to "Consent Form."}

\item{filename}{The name of the file containing the consent form text. This must be an HTML file stored in the app's \code{www} directory. If no filename is given, a generic consent text is used.}
}
\value{
A user interface that displays a custom consent form from a separate file, buttons for the participant to consent or decline, and (optionally) an opportunity for the participant to explicitly agree to audio recording.
}
\description{
Shiny UI function for displaying a participant consent form
}
\examples{
if (interactive()) {
ui <- shiny::fluidPage(
  consentUI(id = "cons")
)

server <- function(input, output, session) {
 consentServer(id = "cons")
}
shiny::shinyApp(ui = ui, server = server)
}

}
\seealso{
Must be used with \code{\link{consentServer}}. If \code{cons2rec = TRUE}, consider using \code{\link{recorderUI}} and \code{\link{recorderServer}}.

Other Consent form module: 
\code{\link{consentServer}()}
}
\concept{Consent form module}
