% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demographics}
\alias{demographics}
\title{Basic Demographic Survey Questions.}
\format{
A data frame with 7 rows (corresponding to 7 survey questions) and 5 variables:
\describe{
\item{id}{A unique input ID for each question}
\item{label}{Labels (i.e., the questions) that will be displayed to participants}
\item{priority}{Specifies whether an answer to a given question is required or not.}
\item{type}{Type of shiny input function to use for each question.}
\item{options}{Contains the list of options for all non-textInput questions.}

}
}
\usage{
demographics
}
\description{
A data frame containing a correctly formatted basic demographic questionnaire, for use with the \verb{speechcollectr::survey*()} functions.
}
\keyword{datasets}
