% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalTranscServer.R
\name{evalTranscServer}
\alias{evalTranscServer}
\title{Evaluate Transcription and Give Feedback in a Shiny App}
\usage{
evalTranscServer(
  filename = NULL,
  text,
  correct,
  alphaOnly = TRUE,
  matchCase = FALSE,
  attempts = NULL,
  counter = 1,
  rmIncorrect = TRUE
)
}
\arguments{
\item{filename}{An RDS file output from \code{transcribeServer()} OR an RDS file containing a single character string representing the entered transcription.}

\item{text}{A character string to test against the correct transcription. Not necessary if \code{filename} is supplied.}

\item{correct}{A character string of the correct transcription to which \code{text} or \code{filename} will be compared.}

\item{alphaOnly}{Boolean. Should the transcription be compared only in alphabetic characters entered, ignoring any spaces, punctuation, or digits? Defaults to TRUE.}

\item{matchCase}{Boolean. Should the comparison be case-sensitive? Defaults to FALSE.}

\item{attempts}{Integer. How many attempts does the user get to enter a correct transcription?}

\item{counter}{Integer. Usually a shiny reactive value. Number of attempts the participant has completed.}

\item{rmIncorrect}{Boolean. Should files containing incorrect transcriptions be kept if the participant has attempts remaining? Defaults to TRUE.}
}
\value{
A Boolean value indicating whether the transcription in \code{filename} matches the value given in \code{correct} (=1) or not (=0). This function will also return feedback to the user in dialog boxes, indicating success (if a correct transcription was entered); a warning (upon entry of an incorrect transcription, informing participants of how many attempts remain); or an error alert (indicating an incorrect transcription when all attempts have been used).
}
\description{
Evaluate Transcription and Give Feedback in a Shiny App
}
\details{
Though transcriptions can be evaluated the basic transcribeServer() function, this function allows users more flexibility by letting them choose when they want to check the transcription and whether it should be checked for case and non-alphanumeric characters.
}
\note{
The parameter \code{rmIncorrect} removes incorrect transcriptions BEFORE the participants' final attempt. This means that the final transcription the participant enters will never be deleted. This parameter was added because, without it, the program will evaluate the current attempt and all previous attempts that have been saved under a given filename, returning numerous feedback screens any time more than one attempt is completed.
}
\examples{
data("samp_wav")
tuneR::writeWave(samp_wav, "sample.wav")
wwwPrep(from = "sample.wav")

if (interactive()) {
 library(shiny)
 library(speechcollectr)
 library(shinyjs)
 ui <- fluidPage(
  actionButton("begin", "Begin"),
  transcribeUI(id = "transcribe"),
  h5("Click 'Begin' each time you want to transcribe. The recording will not change from trial to trial."),
  h5("To demonstrate the transcription evaluation procedure, we'll use a different 'correct transcritiption' for each trial."),
  h5("The correct answers for the first 4 trials are: 'a', 'ab', 'abc', 'abcd' (in that order).")
 )


 server <- function(input, output, session) {
  rvs <- reactiveValues(n = 1, attempt = 1)
  answers <- c("a", "ab", "abc", "abcd")

  observeEvent(input$begin, {
    disable("begin")
    transcribeServer(id="transcribe",
                     audioFile = "sample.wav",
                     n_play = 4,
                     outFile = paste0("sample", rvs$n, ".rds"),
                     result = "hide")
  })

  observeEvent(input[["transcribe-submit"]], {
    enable("begin")

    rvs$is_correct <- evalTranscServer(filename = paste0("sample", rvs$n, ".rds"),
                                       counter = rvs$attempt,
                                       attempts = 2,
                                       matchCase = TRUE,
                                       correct = answers[rvs$n])

    shiny::req(!is.null(rvs$is_correct))
    if (rvs$is_correct == 1) {
      rvs$n <- rvs$n + 1
    } else {
      rvs$attempt <- rvs$attempt + 1
    }
  })

 }
 shinyApp(ui = ui, server = server)
}
}
