% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWav.R
\name{evalWav}
\alias{evalWav}
\title{Check a WAV file's quality}
\usage{
evalWav(
  wave,
  utteranceTG = NULL,
  transcription = NULL,
  trim = NULL,
  plotOsc = NULL
)
}
\arguments{
\item{wave}{Required. Either a \code{tuneR::Wave} object or a valid file path to an existing wav file.}

\item{utteranceTG}{If not \code{NULL} (the default), a filename for the Praat TextGrid to be created. Must have the extension \code{.TextGrid}.}

\item{transcription}{If not \code{NULL} (the default), a character vector indicating the transcription of the utterance.}

\item{trim}{If not \code{NULL} (the default), a filename for the wav file to be created from the output of findSpeech() and the original .wav denoted with argument \code{wave}.}

\item{plotOsc}{If not \code{NULL} (the default), a character vector that will server as the title of the wave form plot (recommended = the name of the object in \code{wave}).}
}
\value{
A list with 7 elements: the markers for the beginning and end of the utterance in samples, the sample rate of the original file, the SNR (dB), a logical value indicating whether 1\% of samples in the original wave file were clipped, a tuneR::Wave object for the trimmed file (if \code{trimmed != NULL}), the text of the Praat TextGrid (if \code{utteranceTG != NULL}), and the plot of the wave form beginning and end markers (if \code{plot != NULL}).
}
\description{
Check a WAV file's quality
}
\examples{
data("samp_wav")
eval <- evalWav(samp_wav, plotOsc = "samp_wav")
print(eval$plot)
}
