% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNR.R
\name{getSNR}
\alias{getSNR}
\title{Get the signal-to-noise ratio of a recording containing a single utterance and some leading or trailing noise}
\usage{
getSNR(wave, begin, end)
}
\arguments{
\item{wave}{Required. Either a \code{tuneR::Wave} object or a valid file path to an existing wav file.}

\item{begin}{Integer. The first sample in the \code{wave} to be considered part of the speech signal. If \code{NULL} (the default), will use the automatically detected value output by findSpeech(wave).}

\item{end}{Integer. The last sample in the \code{wave} to be considered part of the speech signal. If \code{NULL} (the default), will use the automatically detected value output by findSpeech(wave).}
}
\value{
The signal-to-noise ratio of the recording in dB.
}
\description{
Get the signal-to-noise ratio of a recording containing a single utterance and some leading or trailing noise
}
\examples{
data("samp_wav")
fs <- findSpeech(samp_wav)
snr <- getSNR(samp_wav, fs$begin_s, fs$end_s)
snr
}
