% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headphoneTestUI.R
\name{headphoneTestUI}
\alias{headphoneTestUI}
\title{UI function for the Headphone Test Module}
\usage{
headphoneTestUI(
  id = "headphone_test",
  type = c("huggins", "antiphase"),
  n_trials = 6
)
}
\arguments{
\item{id}{The Id of this module. Must match the Id of \code{\link{headphoneTestServer}}. Defaults to "headphone_test".}

\item{type}{Whether to use the Huggins Pitch headphone screen (see Milne et al., 2020) or the Antiphase headphone screen (Woods et al., 2017). Must be one of "huggins" or "antiphase" and must match the value given in \code{headphoneTestServer()}.}

\item{n_trials}{Integer. How many trials should be presented to each participant? Defaults to 6. Must match the n_trials argument of \code{\link{headphoneTestServer}}}
}
\value{
Creates the user interface for a headphone screen. Must be used with \code{headphoneTestServer()} in the server function of the app.
}
\description{
Create the user interface for a headphone screen using Huggins pitch or Antiphase phenomena.
}
\examples{
## First use www_create() to get the data for the type of headphone screen you want.
## NOTE: Do NOT put this command in your UI!
## Run it once before testing your app to create the www folder.
www_create(HugginsPitchScreen = TRUE)

# Now build the app.
if (interactive()) {
 ui <- fluidPage(
        headphoneTestUI(id = "headphone_test", type = "huggins")
     )

 server <- function(input, output, session) {
                           headphoneTestServer(id = "headphone_test", type = "huggins",
                                                n_trials = 6, threshold = 4, n_attempts = 2)
  }
shinyApp(ui = ui, server = server)
}
}
\references{
Milne, A.E., Bianco, R., Poole, K.C., Zhao, S., Oxenham, A.J., Billing, A.J., & Chait, M. 2020. An online headphone screening test based on dichotic pitch.\emph{Behavior Research Methods} 53, 1551-1562 (2021).
Woods, K.J.P., Siegel, M.H., Traer, J., & McDermott, J.H. 2017. Headphone screening to facilitate web-based auditory experiments. \emph{Attention, Perception, & Psychophysics} 79, 2064-2072 (2017).
}
\seealso{
\code{\link{www_create}}, \code{\link{headphoneTestUI}}

Other headphone screen module: 
\code{\link{headphoneTestServer}()}
}
\concept{headphone screen module}
