% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middle.R
\name{middle}
\alias{middle}
\title{Vertically align content in the middle of the page}
\usage{
middle(..., align = "center")
}
\arguments{
\item{...}{Any ui elements}

\item{align}{One of "center" (default), "left", or "right", describing the horizontal alignment of the content.}
}
\value{
A user interface with the content wrapped by the function appearing in the middle of the screen.
}
\description{
Vertically align content in the middle of the page
}
\examples{
library(shiny)
if (interactive()) {
ui <- fluidPage(
  middle(fluidRow(actionButton("test1", "Test1")),
         fluidRow(actionButton("test2", "Test2"))),
  bottom("this is at the bottom", align = "right")
)

server <- function(input, output, session) {
}
shinyApp(ui, server)
}
}
