% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauseBttn.R
\name{pauseBttn}
\alias{pauseBttn}
\title{Pause Button for Audio Playback}
\usage{
pauseBttn(inputId, audioId, label = "", inline = FALSE, icon = "pause")
}
\arguments{
\item{inputId}{The input ID that will be added to the list of other inputs in a shiny app.}

\item{audioId}{The input ID associated with an audio tag called with \code{htmltools::tags$audio()}.}

\item{label}{An optional text label to add to the button. Defaults to blank. NOTE: If icon is \code{NULL}, this variable should be given a value explicitly.}

\item{inline}{Boolean. If FALSE (the default), the oause button will be on its own line in the UI (i.e., wrapped in a \code{div} tag). If TRUE, the button will be on a line with the preceeding shiny UI object (i.e., wrapped in a \code{span} tag).}

\item{icon}{An optional \code{shiny::icon()} to add to the button. Defaults to the font-awesome "pause" icon. NOTE: If no text label is present, icon cannot be null.}
}
\value{
On the server side of a shiny app, this function returns an integer of class "shinyActionButtonValue", which means it can be used in all the ways an ordinary shiny action button can (e.g., you can add functions to be called when the button is pressed with \code{shiny::observeEvent()} and \code{shiny::eventReactive}).
}
\description{
Add an action button to a \code{shiny} user interface with a built-in function of pausing the audio currently playing.
}
\examples{

}
