% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qualifications}
\alias{qualifications}
\title{Questions for checking participant qualifications.}
\format{
A data frame with 7 rows (corresponding to 7 survey questions) and 5 variables:
\describe{
\item{id}{A unique input ID for each question}
\item{label}{Labels (i.e., the questions) that will be displayed to participants}
\item{priority}{Specifies whether an answer to a given question is required or not.}
\item{error}{If an item is required, specifies the error message that will be displayed to participants if answered with FALSE.}

}
}
\usage{
qualifications
}
\description{
A data frame containing a correctly formatted set of questions, for use with the \verb{speechcollectr::check*()} functions.
}
\keyword{datasets}
