% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankPlayer.R
\name{rankPlayer}
\alias{rankPlayer}
\title{Get current player's rank among group of past players}
\usage{
rankPlayer(
  score,
  playerId = NULL,
  rankFile,
  score_col = "score",
  id_col = "pin",
  rank_col = "rank"
)
}
\arguments{
\item{score}{The player's current score.}

\item{playerId}{The player's ID number. MUST be unique for each player! Use \code{\link{pinGen}}.}

\item{rankFile}{A file path to a new or existing .rds or .csv file of previous players' scores. Must contain 3 columns described by the remaining arguments.}

\item{score_col}{The name of the column in \code{rankFile} where the past players' scores are stored. Defaults to "score".}

\item{id_col}{The name of the column in \code{rankFile} where the past players' unique ID numbers are stored. Defaults to "pin".}

\item{rank_col}{The name of the column in \code{rankFile} where the past players' ranks are stored. Defaults to "rank".}
}
\value{
A list with 2 values: (1) rank = the current players rank (integer) and (2) out_of = the total number of players so far.
}
\description{
Get current player's rank among group of past players
}
\examples{
# Erase the out file, "sample_rank.rds" after you run this example,
# ...so you won't get the error of multiple "players" with the same ID next time you run the example!
if (interactive()) {
  library(shiny)
  library(shinyalert)
  library(shinyjs)
  ui <- fluidPage(
    useShinyalert(),
    actionButton("ok", "Click to add to your score!"),
    textOutput("score_txt"),
    actionButton("rank", "Show your rank when you're ready.")
  )

  server <- function(input, output, session) {
    rv <- reactiveValues(score = 0, pin = 1)
    observeEvent(input$ok, {
      rv$score <- rv$score + sample(c(1:9), 1)
      output$score_txt <- renderText({
        paste0("Your current score: ",rv$score)
        })
      enable("rank")
    })

    observeEvent(input$rank, {
      disable("rank")
      rv$pin <- rv$pin+1
      rank <- rankPlayer(score = rv$score,
                         playerId = rv$pin,
                         rankFile = "sample_rank.rds")
      text <- paste0("You rank #", rank$rank,
                     " out of ", rank$out_of, " player(s).")
      shinyalert(title = "Congratulations!",
                 type = "success",
                 text = text)
    })
  }
  shinyApp(ui = ui, server = server)
}
}
