% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorderServer.R
\name{recorderServer}
\alias{recorderServer}
\title{Server function for Recording User Audio}
\usage{
recorderServer(id = "recorder", folder = ".", filename)
}
\arguments{
\item{id}{The input ID associated with the recorder module. Must be the same as the id of \code{recorderUI()}.}

\item{folder}{Character. Where to store the audio file. Defaults to the path returned by \code{getwd()}}

\item{filename}{Required. Character. The name of the file to be recorded. Do not specify a directory here. Instead, do that with the \code{folder} argument.}
}
\value{
Saves a single-channel, 16-bit, 44.1 kHz sample rate WAV file with the specified \code{filename} in a directory specified by \code{folder}.
}
\description{
The server function for recording user audio enables the 'stop' button after the user begins recording and checks to make sure the user has given the website permission to record audio in their browser. Requires the UI \code{\link{recorderUI}}.
}
\examples{

}
\seealso{
Must be used with \code{\link{recorderUI}}.

Other Audio recording module: 
\code{\link{recorderUI}()}
}
\concept{Audio recording module}
