% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorderUI.R
\name{recorderUI}
\alias{recorderUI}
\title{Create UI to Record User Audio}
\usage{
recorderUI(
  id = "recorder",
  startText = "RECORD",
  startTextCol = "white",
  startFillCol = "#228833",
  stopText = "STOP",
  stopTextCol = "black",
  stopFillCol = "#EE6677",
  startInline = FALSE,
  stopInline = TRUE
)
}
\arguments{
\item{id}{The input ID associated with the recorder module. Must be the same as the id of \code{recorderServer()}.}

\item{startText}{Character. The text label for the "start recording" button. Defaults to "RECORD".}

\item{startTextCol}{A valid color name in R or a hexidecimal color code denoting the color of the start button text. Defaults to "white".}

\item{startFillCol}{A valid color name in R or a hexidecimal color code denoting the color of the start button text. Defaults to the "green" hex code from Paul Tol's colorblind-safe \emph{bright} qualitative color scheme.}

\item{stopText}{Character. The text label for the "stop recording" button. Defaults to "STOP".}

\item{stopTextCol}{A valid color name in R or a hexidecimal color code denoting the color of the start button text. Defaults to "black".}

\item{stopFillCol}{A valid color name in R or a hexidecimal color code denoting the color of the start button text. Defaults to Defaults to the "red" hex code from Paul Tol's colorblind-safe \emph{bright} qualitative color scheme..}

\item{startInline}{Boolean. Should the 'start recording' button be on the same line as the preceding UI element?}

\item{stopInline}{Boolean. Should the 'stop recording' button be placed on the same line as the 'start recording' button?}
}
\value{
A user interface containing buttons to control recording of participant audio.
}
\description{
The user-interface side of the Shiny module to record user audio in Speech Production experiments. A participant chooses when to begin recording and when to stop by clicking the buttons created in this module. Requires the server-side function \code{\link{recorderServer}}.
}
\examples{

}
\seealso{
Must be used with \code{\link{recorderServer}}. Paul Tol's colorblind-safe palettes (the source of the default button colors) can be found at \url{https://personal.sron.nl/~pault/#sec:qualitative}.
}
\concept{Audio recording module}
