% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordUI.R
\name{recordUI}
\alias{recordUI}
\title{Create UI to Record User Audio}
\usage{
recordUI(
  id = "record",
  writtenStim = FALSE,
  startText = "RECORD",
  startTextCol = "white",
  startFillCol = "#228833",
  stopText = "STOP",
  stopTextCol = "black",
  stopFillCol = "#EE6677",
  startInline = FALSE,
  stopInline = TRUE,
  align = "center"
)
}
\arguments{
\item{id}{The input ID associated with the record module. Must be the same as the id of \code{recordServer()}.}

\item{writtenStim}{Boolean. Do you have a written stimulus that you want your participant to read?}

\item{startText}{Character. The text label for the "start recording" button. Defaults to "RECORD".}

\item{startTextCol}{A valid color name in R or a hexidecimal color code denoting the color of the start button text. Defaults to "white".}

\item{startFillCol}{A valid color name in R or a hexidecimal color code denoting the color of the start button text. Defaults to the "green" hex code from Paul Tol's colorblind-safe \emph{bright} qualitative color scheme.}

\item{stopText}{Character. The text label for the "stop recording" button. Defaults to "STOP".}

\item{stopTextCol}{A valid color name in R or a hexidecimal color code denoting the color of the start button text. Defaults to "black".}

\item{stopFillCol}{A valid color name in R or a hexidecimal color code denoting the color of the start button text. Defaults to Defaults to the "red" hex code from Paul Tol's colorblind-safe \emph{bright} qualitative color scheme.}

\item{startInline}{Boolean. Should the 'start recording' button be on the same line as the preceding UI element?}

\item{stopInline}{Boolean. Should the 'stop recording' button be placed on the same line as the 'start recording' button?}

\item{align}{One of 'left', 'center', or 'right'. Should the elements in this UI be left-, center-, or right-aligned?}
}
\value{
A user interface (initially hidden; see examples below or \code{\link{recordServer}} for details on how to trigger the appearance of this interface) containing buttons to control recording of participant audio.
}
\description{
The user-interface side of the Shiny module to record user audio in Speech Production experiments. A participant chooses when to begin recording and when to stop by clicking the buttons created in this module. Requires the server-side function \code{\link{recordServer}}.
}
\note{
The "start" and "stop" buttons from this module can be accessed in the server code with the "id" of the module as follows: \code{input[["id-start"]]} or \code{input[["id-stop"]]}
}
\seealso{
Must be used with \code{\link{recordServer}}. Paul Tol's colorblind-safe palettes (the source of the default button colors) can be found at \url{https://personal.sron.nl/~pault/#sec:qualitative}.

Other Audio recording module: 
\code{\link{recordServer}()}
}
\concept{Audio recording module}
