% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samp_wav}
\alias{samp_wav}
\title{Sample audio recording}
\format{
A tuneR::Wave object.
}
\source{
\url{https://librivox.org/the-wisdom-of-father-brown-by-g-k-chesterton/}
}
\usage{
samp_wav
}
\description{
A recording of a single interrogative utterance by a male British English speaker. From the Librivox recording of \emph{The Wisdom of Father Brown} by G.K. Chesterton.
}
\keyword{datasets}
