% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startRec.R
\name{startRec}
\alias{startRec}
\title{Start Recording User Audio in a Shiny App}
\usage{
startRec()
}
\value{
Adds the value "record-ready" to the session's input object. The event \code{input[["rec-ready"]]} event can be used to trigger other actions (like the appearance of some text for the participant to read).
}
\description{
Requests access to the user's microphone and begins audio recording if permission is granted. For use only inside a call to \code{observeEvent()} or \code{bindEvent()} in a Shiny application's server code.
}
\examples{
if (interactive()){
 library(shiny)
 library(shinyjs)

 # Build UI
 ui <- fluidPage(

   ## Get background javascript ready
   useShinyjs(),
   tags$head(
     useRecorder()
   ),

   ## Setup page layout
   div(id = "trialDiv",
       style = "text-align:center;",

       hidden(div(id = "textDiv",
                  h4("Please record yourself reading this sentence aloud:"),
                  h2(textOutput(outputId = "read_this")))),

       ## Create the buttons for controlling the recording
       actionButton(inputId = "start",
                    label = "start"),

       ### Hide the stop button until user clicks start
       hidden(actionButton(inputId = "stop",
                           label = "stop"))
   )
 )

 # The Server function
 server <- function(input, output, session) {
   ## Create an object that will count trials
   rvs <- reactiveValues(trial_n = 0)

   ## When the start button is clicked
   observeEvent(input$start, {

     ### Increase counter by one
     rvs$trial_n <- rvs$trial_n + 1

     ### Start the recording
     startRec()

     ### Disable the start button
     disable("start")

     ### Show the stop button
     delay(500, showElement("stop"))
   })

   observeEvent(input[["rec-ready"]], {
     showElement("textDiv")
     output$read_this <- renderText({paste0("This is recording ",
                                            rvs$trial_n, ".")})
   })

   ## When the user clicks stop
   observeEvent(input$stop, {

     ### Stop recording
     stopRec(filename = paste0("rec", rvs$trial_n, ".wav"))

     ### Enable the start button
     enable("start")

     ### Hide the stop button
     hide("stop")
     hide("textDiv")
   })
 }

 # Run the application
 shinyApp(ui = ui, server = server)
}

}
\seealso{
Should be used with \code{\link{stopRec}}
}
