% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useRecorder.R
\name{useRecorder}
\alias{useRecorder}
\title{Use the recording scripts to record audio}
\usage{
useRecorder(
  frontendJS = "www/rec_frontend.js",
  backendJS = "www/rec_backend.js"
)
}
\arguments{
\item{frontendJS}{Path to the frontend recording script. Downloaded with wwwPrep as "www/rec_frontend.js"}

\item{backendJS}{Path to the backend recording script. Downloaded with wwwPrep as "www/rec_backend.js"}
}
\value{
Adds the JS scripts for audio recording to the ui object.
}
\description{
Use the recording scripts to record audio
}
