% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uttTG.R
\name{uttTG}
\alias{uttTG}
\title{Create a Praat TextGrid marking speech onset and offset}
\usage{
uttTG(wave, outfile, begin = NULL, end = NULL, transcription = NULL)
}
\arguments{
\item{wave}{Required. Either a \code{tuneR::Wave} object or a valid file path to an existing wav file.}

\item{outfile}{If not \code{NULL} (the default), a filename ending in .TextGrid where the output of this function will be stored. If \code{NULL}, the function will return a character vector including the text to put in a TextGrid file.}

\item{begin}{Integer. The first SAMPLE in the \code{wave} to be considered part of the speech signal. If \code{NULL} (the default), will use the automatically detected value output by findSpeech(wave).}

\item{end}{Integer. The last SAMPLE in the \code{wave} to be considered part of the speech signal. If \code{NULL} (the default), will use the automatically detected value output by findSpeech(wave).}

\item{transcription}{Character. The transcription of the segment between the \code{begin} and \code{end} markers.}
}
\value{
A praat TextGrid file.
}
\description{
Create a Praat TextGrid marking speech onset and offset
}
\examples{
data("samp_wav")
uttTG(samp_wav, outfile = "samp_tg.TextGrid",
      transcription = c("why what do you mean asked the detective"))
}
