% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/www_create.R
\name{www_create}
\alias{www_create}
\title{Create the 'www' Directory for a Shiny App}
\usage{
www_create(
  from = NULL,
  is_dir = FALSE,
  path = ".",
  volumeCalibration = FALSE,
  HugginsPitchScreen = FALSE,
  AntiphaseScreen = FALSE
)
}
\arguments{
\item{from}{A character vector of filenames or a single directory name wrapped in c(). Leave \code{NULL} (default) if you only want to add one of the datasets included in the package.}

\item{is_dir}{Boolean. Is 'from' a directory?}

\item{path}{A character vector naming the parent directory of 'www'. Defaults to working directory.}

\item{volumeCalibration}{Boolean. Should \code{speechcollectr::cal_noise()} for testing headphone volume be copied to 'www'?}

\item{HugginsPitchScreen}{Boolean. Should \code{speechcollectr::HugginsPitchData()} for Huggins Pitch Headphone Screen be copied to 'www'?}

\item{AntiphaseScreen}{Boolean. Should \code{speechcollectr::AntiphaseData()} for Antiphase Headphone Screen be copied to 'www'?}
}
\value{
Success messages will be printed in the console as each step towards creating the directory is completed.
}
\description{
Create the 'www' Directory for a Shiny App
}
\examples{

}
