% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateUI.R
\name{rateUI}
\alias{rateUI}
\title{User interface builder for likert scale tasks}
\usage{
rateUI(
  id = "rate",
  type = c("button", "slider"),
  align = "center",
  n_scales = 1,
  scaleFillCol = "white",
  scaleTextCol = "black",
  submitText = "SUBMIT",
  submitFillCol = "white",
  submitTextCol = "black"
)
}
\arguments{
\item{id}{The id of the module. Must be the same as the ID of \code{rateServer()}.}

\item{type}{One of 'button' or 'slider'. Which type of input will the respondents give? Must be the same as in the corresponding call to \code{rateServer()}.}

\item{align}{One of 'left', 'center', or 'right'. Should the elements in this UI be left-, center-, or right-aligned?}

\item{n_scales}{Integer. How many scales should be displayed in the interface? Must be the same value as in the corresponding call to \code{rateServer()}}

\item{scaleFillCol}{A valid hexidecimal code or color name for the scale background. Defaults to white.}

\item{scaleTextCol}{A valid hexidecimal code or color name for the scale text and border. Defaults to black.}

\item{submitText}{Character. Label for the submit button. Defaults to 'SUBMIT'.}

\item{submitFillCol}{A valid hexidecimal code or color name for the submit button background.}

\item{submitTextCol}{A valid hexidecimal code or color name for the submit button text.}
}
\value{
A shiny UI for likert scale tasks.
}
\description{
User interface builder for likert scale tasks
}
\examples{
if (interactive()){
  library(shiny)
  library(shinyjs)
  ui <- fluidPage(
    actionButton("btn", "Click me"),
    rateUI(id = "example",
           type = "button"),
    textOutput("confirmation")
  )

  server <- function(input, output, session) {
  rvs <- reactiveValues(rating = NULL)
    observeEvent(input$btn, {
      disable("btn")

      rvs$rating <- rateServer(id = "example",
                               type = "button",
                               instructions = "What do you think?",
                               answers = c("Strongly disagree", "Disagree",
                                           "Neutral", "Agree", "Strongly agree"),
                               pretext = "The vowels 'aw' and 'ah' sound exactly the same.")
    })

    observeEvent(input[["example-submit"]]{
    enable("btn")
        output$confirmation <- renderText({
          paste0("You selected ", rvs$rating(),".")})

    })
  }
  shinyApp(ui = ui, server = server)
}

# An example with 2 scales....
if (interactive()){
   library(shiny)
   library(shinyjs)
   ui <- fluidPage(
     actionButton("btn", "Click me"),
     rateUI(id = "example",
            type = "button",
            n_scales = 2),
     textOutput("confirmation")
   )

   server <- function(input, output, session) {
     rvs <- reactiveValues(rating = NULL)

     observeEvent(input$btn, {
       disable("btn")

       rvs$rating <- rateServer(id = "example",
                                trigger = NULL,
                                type = "button",
                                instructions = "Finish the sentence:",
                                answers = list(c("Sound completely the same",
                                                 "Sound similar, but not totally alike",
                                                 "Sound pretty different",
                                                 "Sound totally different"),
                                               c("Are produced in the exact same way",
                                                 "Are produced similarly",
                                                 "Are produced pretty distinctly",
                                                 "Are produced in totally distinct ways")),
                                pretext = "The vowels 'aw' and 'ah'...",
                                n_scales = 2,
                                answer_all = TRUE,
                                direction = "vertical",
                                scale_labs = c("perception", "production"))
     })

     observeEvent(input[["example-submit"]], {
       enable("btn")
       output$confirmation <- renderText({
         paste0("You selected '", rvs$rating()[[1]],"' and '",  rvs$rating()[[2]], "'.")})
     })
   }
   shinyApp(ui = ui, server = server)
 }
}
